+++ Alloy/commands/compile/compilerUtils.js
@@ -29,7 +29,8 @@ var RESERVED_ATTRIBUTES = [
 		'formFactor',
 		CONST.BIND_COLLECTION,
 		CONST.BIND_WHERE,
-		CONST.AUTOSTYLE_PROPERTY
+		CONST.AUTOSTYLE_PROPERTY,
+		'ns'
 	],
 	RESERVED_ATTRIBUTES_REQ_INC = [
 		'platform',
@@ -38,7 +39,8 @@ var RESERVED_ATTRIBUTES = [
 		'formFactor',
 		CONST.BIND_COLLECTION,
 		CONST.BIND_WHERE,
-		CONST.AUTOSTYLE_PROPERTY
+		CONST.AUTOSTYLE_PROPERTY,
+		'ns'
 	],
 	RESERVED_EVENT_REGEX =  /^on([A-Z].+)/;
 
@@ -87,6 +89,11 @@ exports.getNodeFullname = function(node) {
 		fullname = ns + '.' + name;
 
 	return fullname;
+};
+
+exports.isNodeForCurrentPlatform = function(node) {
+	return !node.hasAttribute('platform') || !compilerConfig || !compilerConfig.alloyConfig ||
+		node.getAttribute('platform') === compilerConfig.alloyConfig.platform;
 };
 
 exports.getParserArgs = function(node, state, opts) {
@@ -180,7 +187,7 @@ exports.getParserArgs = function(node, state, opts) {
 		var attrName = attr.nodeName;
 		if (_.contains(attrs, attrName)) { return; }
 		var matches = attrName.match(RESERVED_EVENT_REGEX);
-		if (matches !== null) {
+		if (matches !== null && exports.isNodeForCurrentPlatform(node)) {
 			events.push({
 				name: U.lcfirst(matches[1]),
 				value: node.getAttribute(attrName)
@@ -279,7 +286,15 @@ exports.generateNode = function(node, state, defaultId, isTopLevel, isModelOrCol
 	if (state.args) { args.symbol = state.args.symbol || args.symbol; }
 
 	// add to list of top level views, if its top level
-	if (isTopLevel) { code.content += args.symbol + ' && $.addTopLevelView(' + args.symbol + ');'; }
+	if (isTopLevel) {
+		if (state.isProxyProperty) {
+			delete state.isProxyProperty;
+			code.content += state.parent.symbol + ' && $.addProxyProperty("' + state.propertyName +
+				'", ' + state.parent.symbol + ');';
+		} else {
+			code.content += args.symbol + ' && $.addTopLevelView(' + args.symbol + ');';
+		}
+	}
 
 	// handle any model/collection code
 	if (state.modelCode) {
@@ -304,7 +319,7 @@ exports.generateNode = function(node, state, defaultId, isTopLevel, isModelOrCol
 				cb: ev.value,
 				escapedCb: ev.value.replace(/'/g, "\\'"),
 				func: eventFunc
-			};
+			}, postCode;
 
 			// create templates for immediate and deferred event handler creation
 			var theDefer = _.template("__defers['<%= obj %>!<%= ev %>!<%= escapedCb %>']", eventObj);
@@ -319,7 +334,11 @@ exports.generateNode = function(node, state, defaultId, isTopLevel, isModelOrCol
 
 			// add the generated code to the view code and post-controller code respectively
 			code.content += _.template(immediateTemplate, eventObj);
-			exports.postCode += _.template(deferTemplate, eventObj);
+			postCode = _.template(deferTemplate, eventObj);
+			exports.postCode += state.condition ? _.template(codeTemplate, {
+				condition: state.condition,
+				content: postCode
+			}) : postCode;
 		});
 	}
 
@@ -499,21 +518,41 @@ exports.inspectRequireNode = function(node) {
 	};
 };
 
-exports.copyWidgetResources = function(resources, resourceDir, widgetId) {
+exports.copyWidgetResources = function(resources, resourceDir, widgetId, opts) {
+	opts = opts || {};
 	_.each(resources, function(dir) {
 		if (!path.existsSync(dir)) { return; }
+		logger.trace('WIDGET_SRC=' + path.relative(compilerConfig.dir.project, dir));
 		var files = wrench.readdirSyncRecursive(dir);
 		_.each(files, function(file) {
 			var source = path.join(dir, file);
+
+			// make sure the file exists and that it is not filtered
+			if (!fs.existsSync(source) ||
+				(opts.filter && opts.filter.test(file)) ||
+				(opts.exceptions && _.contains(opts.exceptions, file))) {
+				return;
+			}
+
 			if (fs.statSync(source).isFile()) {
-				var destDir = path.join(resourceDir, path.dirname(file), widgetId);
+				var dirname = path.dirname(file);
+				var parts = dirname.split(/[\/\\]/);
+				if (opts.titaniumFolder && parts[0] === opts.titaniumFolder) {
+					dirname = parts.slice(1).join('/');
+				}
+
+				var destDir = path.join(resourceDir, dirname, widgetId);
 				var dest = path.join(destDir, path.basename(file));
 				if (!path.existsSync(destDir)) {
 					wrench.mkdirSyncRecursive(destDir, 0755);
 				}
+
+				logger.trace('Copying ' + file.yellow + ' --> ' +
+					path.relative(compilerConfig.dir.project, dest).yellow + '...');
 				U.copyFileSync(source, dest);
 			}
 		});
+		logger.trace(' ');
 	});
 };
 
@@ -653,7 +692,7 @@ exports.parseConfig = function(file, alloyConfig, o) {
 	_.each(j, function(v,k) {
 		if (!/^(?:env\:|os\:)/.test(k) && k !== 'global') {
 			logger.debug(k + ' = ' + JSON.stringify(v));
-			o[k] = _.extend(o[k] || {}, v);
+			o[k] = v;
 		}
 	});
 
@@ -661,6 +700,8 @@ exports.parseConfig = function(file, alloyConfig, o) {
 		o = _.extend(o, j['global']);
 		o = _.extend(o, j['env:'+alloyConfig.deploytype]);
 		o = _.extend(o, j['os:'+alloyConfig.platform]);
+		o = _.extend(o, j['env:'+alloyConfig.deploytype + ' os:'+alloyConfig.platform]);
+		o = _.extend(o, j['os:'+alloyConfig.platform + ' env:'+alloyConfig.deploytype]);
 	}
 
 	return o;
@@ -685,7 +726,7 @@ exports.loadController = function(file) {
 
 	// get the base controller for this controller
 	code.controller = contents;
-	code.parentControllerName = astController.getBaseController(contents);
+	code.parentControllerName = astController.getBaseController(contents, file);
 
 	return code;
 };
@@ -731,6 +772,7 @@ exports.generateCollectionBindingTemplate = function(args) {
 	// construct code template
 	code += "var " + colVar + "=" + col + ";";
 	code += "function " + handlerFunc + "(e) {";
+	code += "   if (e && e.fromAdapter) { return; }";
 	code += "   var opts = " + handlerFunc + ".opts || {};";
 	code += "	var models = " + whereCode + ";";
 	code += "	var len = models.length;";
